Split profile subcommands
