+++ doc/file_structure.md
@@ -44,7 +44,9 @@ ApplePi is organized around clear TypeScript source boundaries, requirement docu
 │   │       ├── RunCommand.ts
 │   │       ├── SetupCommand.ts
 │   │       ├── SyncCommand.ts
-│   │       └── ProfileCommand.ts
+│   │       ├── ProfileCommand.ts
+│   │       ├── ProfileCreateCommand.ts
+│   │       └── ProfileListCommand.ts
 │   ├── settings/                      # settings loading and merging
 │   │   ├── Settings.ts
 │   │   ├── SettingsLoader.ts
